Dataset Signals context
Under conversion, we have included the calculation that needs to be applied to raw data included in the original data set. For those signals that are incremental (mainly energy values), substraction between last and first value of a given period must be calculated (e.g. energy generated every 15 minutes, one hour or daily)